Updates:
- Add kwargs to some CommonSimUtils functions to allow specifying
allow_hidden_flags
- When using the debug interactions, the results will now also be logged
in the log file.
- Added Yellow and Orange colors for colorizing text with.
- Added the ability to decide whether or not to create a combined.json
file in persistence.
- The version of mods will now be used when logging exception and
messages files.

+++ Scripts/sims4communitylib/persistence/persistence_services/common_folder_persistence_service.py
@@ -28,6 +28,8 @@ class CommonFolderPersistenceService(CommonPersistenceService):
     :type allow_duplicates_in_collections: bool, optional
     :param data_folder_path: Use to specify a custom folder path at the top level for which to save/load data to/from. Default is "Mods/mod_data".
     :type data_folder_path: str, optional
+    :param create_combined_file: If True, after reading through all json files, a combined.json file will be created that will include all other json data. If false, a combined.json file will not be created. Default is false.
+    :type create_combined_file: bool, optional
     """
 
     # noinspection PyMissingOrEmptyDocstring
@@ -40,7 +42,8 @@ class CommonFolderPersistenceService(CommonPersistenceService):
         main_file_name: str='main.json',
         combined_file_name: str='combined.json',
         allow_duplicates_in_collections: bool=False,
-        data_folder_path: str=None
+        data_folder_path: str=None,
+        create_combined_file: bool=False
     ) -> None:
         super().__init__()
         self._main_file_name = main_file_name
@@ -48,6 +51,8 @@ class CommonFolderPersistenceService(CommonPersistenceService):
         self._allow_duplicates_in_collections = allow_duplicates_in_collections
         from sims4communitylib.utils.common_log_utils import CommonLogUtils
         self._data_folder_path = data_folder_path or CommonLogUtils.get_mod_data_location_path()
+        from sims4communitylib.s4cl_configuration import S4CLConfiguration
+        self._create_combined_file = create_combined_file or S4CLConfiguration().create_combined_json
 
     # noinspection PyMissingOrEmptyDocstring
     def load(self, mod_identity: CommonModIdentity, identifier: str=None) -> Dict[str, Any]:
@@ -70,8 +75,11 @@ class CommonFolderPersistenceService(CommonPersistenceService):
 
         file_names.append(self._main_file_name)
 
-        combined_file_path = os.path.join(folder_path, self._combined_file_name)
-        loaded_combined_data: Dict[str, Any] = CommonJSONIOUtils.load_from_file(combined_file_path)
+        if self._create_combined_file:
+            combined_file_path = os.path.join(folder_path, self._combined_file_name)
+            loaded_combined_data: Dict[str, Any] = CommonJSONIOUtils.load_from_file(combined_file_path)
+        else:
+            loaded_combined_data = None
 
         def _on_file_read_failure(file_path: str, ex: Exception):
             log.error('Failed to read file with path {}'.format(file_path), exception=ex)
@@ -88,8 +96,9 @@ class CommonFolderPersistenceService(CommonPersistenceService):
         for (key, val) in loaded_data.items():
             file_names.append(key)
             complete_data = CommonCollectionUtils.merge_dict(complete_data, val, prefer_source_values=True, allow_duplicates_in_collections=self._allow_duplicates_in_collections)
-        if loaded_combined_data is not None:
-            complete_data = CommonCollectionUtils.merge_dict(loaded_combined_data, complete_data, prefer_source_values=True, allow_duplicates_in_collections=self._allow_duplicates_in_collections)
+        if self._create_combined_file:
+            if loaded_combined_data is not None:
+                complete_data = CommonCollectionUtils.merge_dict(loaded_combined_data, complete_data, prefer_source_values=True, allow_duplicates_in_collections=self._allow_duplicates_in_collections)
         complete_data = CommonCollectionUtils.merge_dict(complete_data, loaded_main_data, prefer_source_values=True, allow_duplicates_in_collections=self._allow_duplicates_in_collections)
         log.format_with_message('Done loading data.', mod=mod_identity, folder_path=folder_path, complete_data=complete_data, file_names=file_names)
         complete_data['loaded_file_names'] = file_names
@@ -102,19 +111,21 @@ class CommonFolderPersistenceService(CommonPersistenceService):
         folder_path = self._folder_path(mod_identity, identifier=identifier)
         data_to_save = data.copy()
 
-        file_path = os.path.join(folder_path, self._combined_file_name)
-        log.format_with_message('Loading data.', mod=mod_identity, file_path=file_path)
-        if 'loaded_file_names' in data_to_save:
-            del data_to_save['loaded_file_names']
+        if self._create_combined_file:
+            file_path = os.path.join(folder_path, self._combined_file_name)
+            log.format_with_message('Loading data.', mod=mod_identity, file_path=file_path)
+            if 'loaded_file_names' in data_to_save:
+                del data_to_save['loaded_file_names']
 
-        os.makedirs(folder_path, exist_ok=True)
-        if os.path.exists(file_path):
-            log.debug('File existed already, removing the existing one.')
-            os.remove(file_path)
+            os.makedirs(folder_path, exist_ok=True)
+            if os.path.exists(file_path):
+                log.debug('File existed already, removing the existing one.')
+                os.remove(file_path)
 
-        result = CommonJSONIOUtils.write_to_file(file_path, data_to_save)
-        log.format_with_message('Done saving data.', file_path=file_path)
-        return result
+            result = CommonJSONIOUtils.write_to_file(file_path, data_to_save)
+            log.format_with_message('Done saving data.', file_path=file_path)
+            return result
+        return True
 
     # noinspection PyMissingOrEmptyDocstring
     def remove(self, mod_identity: CommonModIdentity, identifier: str=None) -> bool:
@@ -122,15 +133,17 @@ class CommonFolderPersistenceService(CommonPersistenceService):
         log = CommonLogRegistry().register_log(mod_identity, '{}_{}'.format(mod_identity.base_namespace, self.log_identifier))
         folder_path = self._folder_path(mod_identity, identifier=identifier)
 
-        file_path = os.path.join(folder_path, self._combined_file_name)
-        log.format_with_message('Removing data.', mod=mod_identity, file_path=file_path)
+        if self._create_combined_file:
+            file_path = os.path.join(folder_path, self._combined_file_name)
+            log.format_with_message('Removing data.', mod=mod_identity, file_path=file_path)
 
-        if os.path.exists(file_path):
-            log.debug('Data existed, removing it.')
-            os.remove(file_path)
+            if os.path.exists(file_path):
+                log.debug('Data existed, removing it.')
+                os.remove(file_path)
 
-        log.format_with_message('Data deleted successfully.', file_path=file_path)
-        return not os.path.exists(file_path)
+            log.format_with_message('Data deleted successfully.', file_path=file_path)
+            return not os.path.exists(file_path)
+        return True
 
     def _folder_path(self, mod_identity: CommonModIdentity, identifier: str=None) -> str:
         folder_path = os.path.join(self._data_folder_path, mod_identity.base_namespace.lower())
